Mapsui.UI.Maui: move all Android-specific parts of the MapControl into a platform-specific file
